Reorganise use case factories
