+++ src/app/edit_routine/AddExerciseUseCaseFactory.java
@@ -1,4 +1,4 @@
-package app;
+package app.edit_routine;
 
 public class AddExerciseUseCaseFactory {
