+++ src/app/edit_routine/RenameRoutineUseCaseFactory.java
@@ -1,0 +1,4 @@
+package app.edit_routine;
+
+public class RenameRoutineUseCaseFactory {
+}
